taint/test/mochitest/test_script_sinks.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test Cookie Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script src="/tests/taint/test/mochitest/taint_helper.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let string_content = "hello";
      let sink_names = [
          "script.src",
          "script.text",
          "script.innerHTML",
          "script.textContent",
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
        if(i == 0) {
          SimpleTest.is(report.detail.str, "url", "Check sink string content");
        } else {
          SimpleTest.is(report.detail.str, "console.log();", "Check sink string content");
        }
          let flow = report.detail.str.taint[0].flow;
          SimpleTest.is(flow[2].operation, sink_names[i], `${sink_names[i]} sink test`);

          i += 1;
          if (i >= sink_names.length) {
              SimpleTest.finish();
          }
      }, false);

      function setScriptProperty(text, f) {
        let script = document.createElement("script");
        script.id = "tempscript";
        f(script, text);
        document.body.appendChild(script);
        document.body.removeChild(script);
      }

      function startTest() {
          let tainted = String.tainted("console.log();");
          setScriptProperty(String.tainted("url"), (s,t) => s.src = t);
          setScriptProperty(tainted, (s,t) => s.text = t);
          setScriptProperty(tainted, (s,t) => s.innerHTML = t);
          setScriptProperty(tainted, (s,t) => s.textContent = t);
      }
      
    </script>

  </head>

  <body onload="startTest();">
  </body>

</html>
